Replace react-scripts with Vite
Migrate the JavaScript build system from react-scripts to Vite. react-scripts is
part of create-react-app, which is no longer supported.

Contentful provide templates which use Vite, so that looks like a good choice
for a new build system. The changes in this commit are based on the "React +
Vite" Contentful app template.